Software work
I2C4 DMA opened.
LTDC CLK frequency decreased for TFT screen.
All communication callback moved to the communucation.c source file.
Bootloader mode switch pin added.
lvgl demos and examples closed.
some function names changed.

+++ Software/H723_InterfaceMaster/Core/Src/lm75b_temp_ic.c
@@ -118,7 +118,7 @@ LM75B_Status_e lm75b_init(void) {
   I2C_DeviceStatus_e status;
   status = IM_I2C_bus_scanner(I2C_ADDRESS);
 
-  if (LM75B_OK != status) {
+  if ((uint8_t)LM75B_OK != status) {
     return LM75B_ERROR;
   }
